Potentiometer + Continuous servo motor
This example will make a servo rotate if the value of the potentiometer goes below a set value. With this, three LED's will start to light up the more you twist the potentiometer. This example was used in a project to simulate wind energy and is useful for younger students.